refactor: move BackgroundAction types to shared dir
